fix: nomenclatures
