+++ .gitignore
@@ -3,3 +3,5 @@ node_modules
 dev
 package-lock.json
 temp.*
+coverage
+junit.xml
